requirements.txt
feedparser>=6.0.10
google-generativeai>=0.3.0
google-cloud-texttospeech>=2.14.0
feedgen>=0.9.0
pytest>=7.0.0
PyYAML>=6.0
tenacity>=8.0.0
